Remove send_msg, recv_msg from stream auth interceptor
Refreshing the authentication token for streaming is challenging due to the stream API.

The server streaming needs the buffered binary from the application state. Therefore, we cannot easily replace the old stream.

There are only two bidirectional streaming calls in the etcd API: lease keep alive and watch. The lease keep alive API doesn’t need an auth token. The watch API is too flexible, allowing users to send a single watch, multiple watches in a watch stream, or even create a watch within an existing watch stream. This flexibility makes it hard for us to adapt to refreshing tokens.